requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
supabase
python-dotenv
